# tools-api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenCodeForge Tools API
  version: 1.0.0
  description: |
    Sandboxed file/git/shell access scoped to a single workspace. Designed to
    be registered as a Tool Server in Open WebUI so the LLM can read, edit
    and execute code inside ./workspace, similar to Claude Code.
servers:
  - url: http://localhost:8088
    description: Local default
  - url: http://tools-api:8088
    description: From inside the docker compose network (open-webui, code-server)
paths:
...
  /read:
    get:
      operationId: readFile
      summary: Read a text file
      parameters:
        - in: query
          name: path
          required: true
          schema: { type: string }
      responses:
        "200": { description: File contents }
        "404": { description: Not found }
...